+++ packages/core/src/payment/strategies/stripe-upe/stripe-upe-payment-strategy.ts
@@ -58,7 +58,6 @@ export default class StripeUPEPaymentStrategy implements PaymentStrategy {
     private _stripeElements?: StripeElements;
     private _isMounted = false;
     private _unsubscribe?: () => void;
-    private _isDeinitialize?: boolean;
 
     constructor(
         private _store: CheckoutStore,
@@ -82,8 +81,6 @@ export default class StripeUPEPaymentStrategy implements PaymentStrategy {
                 'Unable to initialize payment because "gatewayId" argument is not provided.',
             );
         }
-
-        this._isDeinitialize = false;
 
         this._loadStripeElement(stripeupe, gatewayId, methodId).catch((error) =>
             stripeupe.onError?.(error),
@@ -113,8 +110,7 @@ export default class StripeUPEPaymentStrategy implements PaymentStrategy {
                         stripeupe.onError?.(error);
                     } else if (!this._isMounted) {
                         await this._stripeElements?.fetchUpdates();
-                        payment.mount(`#${stripeupe.containerId}`);
-                        this._isMounted = true;
+                        this._mountElement(payment, stripeupe.containerId);
                     }
                 }
             },
@@ -219,7 +215,6 @@ export default class StripeUPEPaymentStrategy implements PaymentStrategy {
 
         this._stripeElements?.getElement(StripeElementType.PAYMENT)?.unmount();
         this._isMounted = false;
-        this._isDeinitialize = true;
 
         return Promise.resolve(this._store.getState());
     }
@@ -436,16 +431,7 @@ export default class StripeUPEPaymentStrategy implements PaymentStrategy {
                 },
             });
 
-        try {
-            stripeElement.mount(`#${containerId}`);
-            this._isMounted = true;
-        } catch (error) {
-            if (!this._isDeinitialize) {
-                throw new InvalidArgumentError(
-                    'Unable to mount Stripe component without valid container ID.',
-                );
-            }
-        }
+        this._mountElement(stripeElement, containerId);
 
         stripeElement.on('ready', () => {
             render();
@@ -660,5 +646,14 @@ export default class StripeUPEPaymentStrategy implements PaymentStrategy {
             stripePublishableKey,
             stripeConnectedAccount,
         );
+    }
+
+    private _mountElement(stripeElement: StripeElement, containerId: string): void {
+        if (!document.getElementById(containerId)) {
+            return;
+        }
+
+        stripeElement.mount(`#${containerId}`);
+        this._isMounted = true;
     }
 }